requirements.txt
holidays
tweepy
openai
pytz
python-dotenv